Fix FK type mismatch: user_id must be int not bigint
auth_user.id is int (Django default AutoField), so all FK columns
referencing it must also be int. Fixes ProgrammingError on startup
when creating ips_savedproposal.

+++ app/ips/migrations/0005_savedproposal.py
@@ -43,7 +43,7 @@ def create_savedproposal_table(apps, schema_editor):
         BEGIN
             CREATE TABLE ips_savedproposal (
                 id                    bigint        IDENTITY(1,1) NOT NULL,
-                user_id               bigint                      NOT NULL,
+                user_id               int                         NOT NULL,
                 label                 nvarchar(200)               NOT NULL,
                 data                  nvarchar(max)               NOT NULL,
                 risk_profile_override nvarchar(100)               NOT NULL DEFAULT '',
